// frontend/StockManager/generated-sources/openapi/models/PortfolioDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PortfolioItemDto } from './PortfolioItemDto';
import {
    PortfolioItemDtoFromJSON,
    PortfolioItemDtoFromJSONTyped,
    PortfolioItemDtoToJSON,
    PortfolioItemDtoToJSONTyped,
} from './PortfolioItemDto';

/**
 * 
 * @export
 * @interface PortfolioDto
 */
export interface PortfolioDto {
    /**
     * 
     * @type {number}
     * @memberof PortfolioDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PortfolioDto
     */
    name?: string | null;
    /**
     * 
     * @type {Array<PortfolioItemDto>}
     * @memberof PortfolioDto
     */
    portfolioItems?: Array<PortfolioItemDto> | null;
    /**
     * 
     * @type {number}
     * @memberof PortfolioDto
     */
    value?: number | null;
    /**
     * 
     * @type {number}
     * @memberof PortfolioDto
     */
    profitLoss?: number | null;
}

/**
 * Check if a given object implements the PortfolioDto interface.
 */
export function instanceOfPortfolioDto(value: object): value is PortfolioDto {
    return true;
}

export function PortfolioDtoFromJSON(json: any): PortfolioDto {
    return PortfolioDtoFromJSONTyped(json, false);
}

export function PortfolioDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PortfolioDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'portfolioItems': json['portfolioItems'] == null ? undefined : ((json['portfolioItems'] as Array<any>).map(PortfolioItemDtoFromJSON)),
        'value': json['value'] == null ? undefined : json['value'],
        'profitLoss': json['profitLoss'] == null ? undefined : json['profitLoss'],
    };
}

export function PortfolioDtoToJSON(json: any): PortfolioDto {
    return PortfolioDtoToJSONTyped(json, false);
}

export function PortfolioDtoToJSONTyped(value?: PortfolioDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'portfolioItems': value['portfolioItems'] == null ? undefined : ((value['portfolioItems'] as Array<any>).map(PortfolioItemDtoToJSON)),
        'value': value['value'],
        'profitLoss': value['profitLoss'],
    };
}